2.0b2

+++ build.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# IMPORTANT - update version in setup.py and doc-src/conf.py
+# IMPORTANT - update version in setup.py
 
 # this generates a new release, but does not register anything with pypi
 # or upload files to google code
@@ -16,8 +16,11 @@ sed -i -e "s/version = .*/version = '$VERSION'/" doc-src/conf.py
 
 sed -i -e "s/__version__ = .*/__version__ = '$RELEASE'/" src/lepl/__init__.py
 
-echo "update version in setup.py and doc-src/conf.py"
-exit
+if [ -e "dist/LEPL-$RELEASE.tar.gz" ]; then
+  echo "update version in setup.py and doc-src/conf.py"
+  echo "current release: $RELEASE"
+  exit
+fi
 
 rm MANIFEST.in
 find . -exec echo "exclude {}" \; | sed -e "s/\.\///" >> MANIFEST.in
